requirements.txt
boto3==1.16.63
requests
RPi.GPIO==0.6.5
pyaudio==0.2.11